rename and fix lint
