+++ fibber/benchmark/benchmark.py
@@ -30,6 +30,7 @@ parser.add_argument("--bert_clf_steps", type=int, default=20000)
 
 RandomStrategy.add_parser_args(parser)
 IdentityStrategy.add_parser_args(parser)
+TextFoolerStrategy.add_parser_args(parser)
 BertSamplingStrategy.add_parser_args(parser)
 
 G_EXP_NAME = None
@@ -66,7 +67,6 @@ def get_strategy(arg_dict, dataset_name, strategy_name, strategy_gpu_id,
             arg_dict, dataset_name, strategy_gpu_id, output_dir, metric_bundle)
     else:
         assert 0
-    dataset_name, strategy_gpu_id, output_dir,
 
 
 def benchmark(arg_dict, dataset_name, trainset, testset, paraphrase_set):
@@ -82,7 +82,6 @@ def benchmark(arg_dict, dataset_name, trainset, testset, paraphrase_set):
     logger.info("Build metric bundle.")
 
     metric_bundle = MetricBundle(
-        # use_glove_semantic_similarity=False,
         use_bert_clf_prediction=True,
         use_gpu_id=arg_dict["use_gpu_id"], gpt2_gpu_id=arg_dict["gpt2_gpu_id"],
         bert_gpu_id=arg_dict["bert_gpu_id"], dataset_name=dataset_name,
